# dr-stone-api/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dr. Stone API",
    "description": "HTTP API for tracked products, collection runs, and historical price analysis.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
...
  "paths": {
...
    "/sources": {
      "get": {
        "tags": ["Sources"],
        "summary": "List canonical sources",
        "operationId": "listSources",
        "responses": {
          "200": {
            "description": "Known sources and whether they are active in the current runtime.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceCatalogResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/search-runs": {
      "get": {
        "tags": ["Search Runs"],
        "summary": "List search runs",
        "operationId": "listSearchRuns",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "description": "Optional UTC calendar date filter in YYYY-MM-DD format.",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of runs to return. Defaults to 40 and is capped at 200.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 40
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search runs and their persisted matched items.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchRunsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/tracked-products": {
      "get": {
        "tags": ["Tracked Products"],
        "summary": "List tracked products",
        "operationId": "listTrackedProducts",
        "parameters": [
          {
            "name": "all",
            "in": "query",
            "description": "Send `1` to include inactive tracked products.",
            "schema": {
              "type": "string",
              "enum": ["1"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tracked products ordered by creation time.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrackedProduct"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Tracked Products"],
        "summary": "Create a tracked product",
        "operationId": "createTrackedProduct",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackedProductCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tracked product created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackedProduct"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/collect-due": {
      "post": {
        "tags": ["Collection"],
        "summary": "Collect all due tracked products",
        "operationId": "collectDueTrackedProducts",
        "responses": {
          "200": {
            "description": "Collection results for every active tracked product.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CollectionResult"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/tracked-products/{trackedProductId}": {
      "parameters": [
        {
          "name": "trackedProductId",
          "in": "path",
          "required": true,
          "description": "Tracked product identifier.",
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "tags": ["Tracked Products"],
        "summary": "Get a tracked product",
        "operationId": "getTrackedProduct",
        "responses": {
          "200": {
            "description": "Tracked product details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackedProduct"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Collection"],
        "summary": "Collect one tracked product",
        "operationId": "collectTrackedProduct",
        "parameters": [
          {
            "name": "action",
            "in": "query",
            "required": true,
            "description": "Must be set to `collect`.",
            "schema": {
              "type": "string",
              "enum": ["collect"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection result for the tracked product.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": ["Tracked Products"],
        "summary": "Replace tracked product fields",
        "operationId": "replaceTrackedProduct",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackedProductUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated tracked product.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackedProduct"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "patch": {
        "tags": ["Tracked Products"],
        "summary": "Update tracked product fields",
        "operationId": "updateTrackedProduct",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackedProductUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated tracked product.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackedProduct"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": ["Tracked Products"],
        "summary": "Delete a tracked product",
        "operationId": "deleteTrackedProduct",
        "responses": {
          "204": {
            "description": "Tracked product deleted."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/tracked-products/{trackedProductId}/history": {
      "get": {
        "tags": ["Price History"],
        "summary": "Get tracked product history",
        "operationId": "getTrackedProductHistory",
        "parameters": [
          {
            "name": "trackedProductId",
            "in": "path",
            "required": true,
            "description": "Tracked product identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return. Defaults to 100 and is capped at 500.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Zero-based offset for pagination.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "start_at",
            "in": "query",
            "description": "Optional start bound. Accepts YYYY-MM-DD or ISO 8601 datetime.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end_at",
            "in": "query",
            "description": "Optional end bound. Accepts YYYY-MM-DD or ISO 8601 datetime.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "in": "query",
            "description": "Optional source filter. Defaults to `all`. Accepts `all` or a valid `source_name`.",
            "schema": {
              "type": "string",
              "default": "all"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated saved history window for a tracked product.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "BadRequest": {
        "description": "Invalid request payload or query parameters.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Requested resource was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Unhandled server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string"
          },
          "error_type": {
            "type": "string"
          }
        }
      },
...
      "Source": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SourceMetadata"
          },
          {
            "type": "object",
            "required": ["active"],
            "properties": {
              "active": {
                "type": "boolean"
              }
            }
          }
        ]
      },
      "SourceCatalogResponse": {
        "type": "object",
        "required": ["sources"],
        "properties": {
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Source"
            }
          }
        }
      },
      "TrackedProduct": {
        "type": "object",
        "required": ["id", "title", "search_terms", "active", "created_at", "updated_at"],
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "search_terms": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "active": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "TrackedProductCreateRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TrackedProductWriteBase"
          },
          {
            "type": "object",
            "required": ["title"],
            "anyOf": [
              {
                "required": ["search_terms"]
              },
              {
                "required": ["search_term"]
              }
            ]
          }
        ]
      },
      "TrackedProductUpdateRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TrackedProductWriteBase"
          }
        ],
        "description": "Any subset of writable tracked product fields."
      },
      "CollectionResult": {
        "type": "object",
        "required": [
          "tracked_product_id",
          "search_run_ids",
          "successful_runs",
          "failed_runs",
          "total_results",
          "matched_results",
          "page_count"
        ],
        "properties": {
          "tracked_product_id": {
            "type": "string"
          },
          "search_run_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "successful_runs": {
            "type": "integer"
          },
          "failed_runs": {
            "type": "integer"
          },
          "total_results": {
            "type": "integer"
          },
          "matched_results": {
            "type": "integer"
          },
          "page_count": {
            "type": "integer"
          }
        }
      },
...
      "SearchRunsResponse": {
        "type": "object",
        "required": ["date", "runs"],
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "runs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchRun"
            }
          }
        }
      }
    }
  }
}